move path id restaraunts method to index
