+++ server/api/index.js
@@ -3,7 +3,6 @@ module.exports = router;
 
 router.use('/users', require('./users'));
 router.use('/path', require('./path'));
-router.use('/scores', require('./scores'));
 router.use('/quiz', require('./quiz'));
 router.use('/game', require('./game'));
 router.use('/favorite', require('./favorite'));
