+++ server/api/path.js
@@ -7,21 +7,7 @@ module.exports = router;
 
 router.get('/:id/restaurants', async (req, res, next) => {
   try {
-    const path_name = req.params.id;
-
-    const path = await Path.findAll({
-      where: { path_name },
-      include: Restaurant,
-      order: [['stage', 'ASC']],
-    });
-
-    const rests = path.map((path) => {
-      let rest = path.restaurant;
-      rest.dataValues.game_type = path.game_type;
-      return rest;
-    });
-
-    res.send(rests);
+    res.send(await Path.getRestaurants(req.params.id));
   } catch (err) {
     next(err);
   }
